Feat: Improve code
