+++ server/constants/style.py
@@ -1,0 +1,3 @@
+container_style = "width: 75%; margin: 0 auto; background: #081223; padding: 20px; border-radius: 8px; color: #ffff !important; text-align: center;"
+anchor_style = "display: inline-block; padding: 10px; background-color: transparent; color: #64FFDA; text-decoration: none; border: 1px solid #64FFDA; border-radius: 5px;"
+anchor_p_style = "margin-top: 35px; color: #ffff;"
